+++ docs/additional_information/changelog.rst
@@ -12,6 +12,7 @@ Added
 - Implemented ``ClusterBasedLocalOutlierFactor`` (CBLOF) anomaly detector.
 - Implemented ``KMeansAnomalyDetector`` anomaly detector.
 - Implemented ``CopulaBasedOutlierDetector`` (COPOD) anomaly detector.
+- Implemented ``RobustScaler`` preprocessor.
 
 Changed
 ^^^^^^^